moved text box to common and updated tile occupy function for player spawn in combat scene
